adding tests & final adjust

+++ lib/core/utils/storage.dart
@@ -1,15 +1,21 @@
+import 'package:carousel_user_story/core/utils/is_test_environment.dart';
 import 'package:carousel_user_story/main.dart';
 
 class Storage {
   static Future<void> clear() async {
-    await MyApp.storage.deleteAll();
+    if (!TestEnvironment.isTestMode()) await MyApp.storage.deleteAll();
   }
 
   static Future<void> setId(String Id) async {
-    await MyApp.storage.write(key: 'Id', value: Id);
+    if (!TestEnvironment.isTestMode()) {
+      await MyApp.storage.write(key: 'Id', value: Id);
+    }
   }
 
   static Future<String?> getId() async {
-    return await MyApp.storage.read(key: 'Id');
+    if (!TestEnvironment.isTestMode()) {
+      return await MyApp.storage.read(key: 'Id');
+    }
+    return "";
   }
 }
